Do the basic static casts

+++ src/object_properties.cpp
@@ -98,7 +98,7 @@ std::string ObjectProperties::dump() const
 	os << ", shaded=" << shaded;
 	os << ", show_on_minimap=" << show_on_minimap;
 	os << ", nametag_scale_z=" << nametag_scale_z;
-	os << ", new_step_up=" << enum_to_string(es_NewStepUp, new_step_up);
+	os << ", new_step_up=" << enum_to_string(es_NewStepUp, static_cast<u8>(new_step_up));
 	return os.str();
 }
 
@@ -225,7 +225,7 @@ void ObjectProperties::serialize(std::ostream &os) const
 		writeU32(os, nametag_fontsize.value());
 
 	writeU8(os, nametag_scale_z);
-	os << serializeString16(enum_to_string(es_NewStepUp, new_step_up));
+	os << serializeString16(enum_to_string(es_NewStepUp, static_cast<u8>(new_step_up)));
 
 	// Add stuff only at the bottom.
 	// Never remove anything, because we don't want new versions of this!
